Paramaterized for GH Org and report location

+++ report.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
 
+GH_ORG=$1
+REPORT="$PWD/$2"
+
+if [ "$2" == "" ]; then
+  echo "Usgae: $0 <GH Org name> <Report output path>"
+  exit -1
+fi
+
 rm -rf /tmp/consc
 mkdir /tmp/consc
 
-echo "Repo, Master, Slave, Blacklist, Whitelist" > ~/dev/consc-lang/report.csv
+echo "Repo, Master, Slave, Blacklist, Whitelist" > $REPORT
 
-REPOS=$(gh repo list quarkusio | awk -F ' ' ' { print $1 }')
+REPOS=$(gh repo list $GH_ORG | awk -F ' ' ' { print $1 }')
 
 for REPO in $REPOS; do
   
@@ -21,5 +29,5 @@ for REPO in $REPOS; do
   WHITELIST_COUNT=$(grep -rni 'whitelist' . | wc -l)
 
 
-  echo "$REPO_NAME,$MASTER_COUNT,$SLAVE_COUNT,$BLACKLIST_COUNT,$WHITELIST_COUNT" >> ~/dev/consc-lang/report.csv
+  echo "$REPO_NAME,$MASTER_COUNT,$SLAVE_COUNT,$BLACKLIST_COUNT,$WHITELIST_COUNT" >> $REPORT
 done
